fix's

+++ tests/user_crud_test.py
@@ -76,17 +76,18 @@ def test_update_user(test_client):
         "user_lastname": fake.last_name()[:50],
     }
     headers = {"Authorization": f"Bearer {access_token}"}
-    response = test_client.put(f"/users/", headers=headers, json=upgrade_data)
+    response = test_client.put(f"/users/{user_id}", headers=headers, json=upgrade_data)
 
     assert response.status_code == 200
 
 
 def test_update_user_not_authorisation(test_client):
+    user_id = signup_data["id"]
     upgrade_data = {
         "user_firstname": fake.first_name()[:50],
         "user_lastname": fake.last_name()[:50],
     }
-    response = test_client.put(f"/users/", json=upgrade_data)
+    response = test_client.put(f"/users/{user_id}", json=upgrade_data)
 
     assert response.status_code == 401
 
@@ -109,6 +110,7 @@ def test_delete_user(test_client):
 
 
 def test_delete_user_not_authorisation(test_client):
-    response = test_client.delete(f"/users/")
+    user_id = signup_data["id"]
+    response = test_client.delete(f"/users/{user_id}")
 
     assert response.status_code == 401
